feat (private): implement support for derivations
