+++ packages/schema-record/rollup.config.mjs
@@ -17,7 +17,7 @@ export default {
   plugins: [
     // These are the modules that users should be able to import from your
     // addon. Anything not listed here may get optimized away.
-    addon.publicEntrypoints(['index.js']),
+    addon.publicEntrypoints(['hooks.js', 'index.js', 'record.js', 'schema.js']),
 
     nodeResolve({ extensions: ['.ts'] }),
     babel({
